draft of API changes
